Add styles to Login form
